requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
pymongo==4.9.1
openai==1.52.2
feedparser==6.0.11
python-dotenv==1.0.1